Added security and ux fixes to chatbot screen
Added a maximum message length for both the text input and the actual request.
Added a rate limit of 1 message per second just in case. This also disables the send button.
Added error alerts for messages that are too long/short.
Added an error alert for when fetching fails.
Changed the main view to be KeyboardAvoiding since SafeAreaView does not work for Android.